🚚 refactor: 포스트페이지 폴더 구조 수정
