+++ src/Router.tsx
@@ -3,16 +3,16 @@ import { Navigate, Route, Routes } from 'react-router';
 import { PATH } from '@/constants/path';
 
 import App from '@/App';
-import AlbumPage from '@/pages/Albums';
-import AlbumEditor from '@/pages/Albums/components/Editor';
-import AlbumList from '@/pages/Albums/components/List';
+import AlbumEditor from '@/pages/AlbumEditor';
+import AlbumList from '@/pages/AlbumList';
+import AlbumPage from '@/pages/AlbumsRouter';
 import Articles from '@/pages/Articles';
 import DailyCheck from '@/pages/DailyCheck';
 import Main from '@/pages/Main';
 import NotFound from '@/pages/NotFound';
-import PostPage from '@/pages/Post';
-import PostEditor from '@/pages/Post/components/Editor';
-import PostList from '@/pages/Post/components/List';
+import PostEditor from '@/pages/PostEditor';
+import PostList from '@/pages/PostList';
+import PostPage from '@/pages/PostRouter';
 import ProjectEditor from '@/pages/ProjectEditor';
 import ProjectItem from '@/pages/ProjectItem';
 import ProjectList from '@/pages/ProjectList';
