+++ src/pages/PostEditor/index.tsx
@@ -4,7 +4,7 @@ import Aside from '@/components/Aside';
 import Button from '@/components/Button';
 import Content from '@/components/Content';
 
-import MarkdownEditor from '../MarkdownEditor';
+import MarkdownEditor from './components/MarkdownEditor';
 import { S } from './style';
 
 /** 포스트 작성 페이지 */
