+++ src/pages/PostList/components/PostingLink/style/index.ts
