+++ src/pages/PostList/index.tsx
@@ -1,7 +1,7 @@
 import Aside from '@/components/Aside';
 import Content from '@/components/Content';
 
-import PostingLink from '../PostingLink';
+import PostingLink from './components/PostingLink';
 
 /** 포스트 목록을 보여주는 페이지 */
 export default function PostList() {
